refactor(all) 패키지 구조 변경
- 프로젝트의 패키지를 view, domain, util로 분리
- UI 클래스는 view로, 도메인 로직은 domain으로, 유틸성 클래스(난수 생성기 등)는 util로 정리
