+++ src/main/java/racingcar/RacingGameApplication.java
@@ -1,5 +1,10 @@
 package racingcar;
 
+import racingcar.domain.DefaultRacingGameRule;
+import racingcar.domain.RacingGame;
+import racingcar.util.RandomNumberGenerator;
+import racingcar.view.RacingGameUI;
+
 public class RacingGameApplication {
 
 	public static void main(String[] args) {
