+++ src/main/java/racingcar/domain/RacingGameRule.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public interface RacingGameRule {
