+++ src/main/java/racingcar/domain/dto/CarDto.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.domain.dto;
 
 public class CarDto {
