+++ src/test/java/racingcar/DefaultRacingGameRuleTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.DefaultRacingGameRule;
 
 @DisplayName("전진 여부를 판단하는 클래스 테스트")
 public class DefaultRacingGameRuleTest {
